mudanças finais 13

+++ menu/models.py
@@ -13,7 +13,7 @@ class Pedido(models.Model):
     mesa = models.IntegerField()
     menu_item = models.ForeignKey(MenuItem, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True)
-    quantidade = models.PositiveBigIntegerField(default=1)
+    quantidade = models.IntegerField(default=1)
     concluido = models.BooleanField(default=False)
     def __str__(self):
         return f"Pedido de {self.menu_item.name} em {self.created_at}"
